chore: setup page migration
