+++ apps/web/pages/auth/setup/index.tsx
@@ -1,15 +1,10 @@
-import type { GetServerSidePropsContext } from "next";
 import { usePathname, useRouter } from "next/navigation";
 import { useState } from "react";
 
 import AdminAppsList from "@calcom/features/apps/AdminAppsList";
-import { getServerSession } from "@calcom/features/auth/lib/getServerSession";
-import { getDeploymentKey } from "@calcom/features/ee/deployment/lib/getDeploymentKey";
 import { APP_NAME } from "@calcom/lib/constants";
 import { useCompatSearchParams } from "@calcom/lib/hooks/useCompatSearchParams";
 import { useLocale } from "@calcom/lib/hooks/useLocale";
-import prisma from "@calcom/prisma";
-import { UserPermissionRole } from "@calcom/prisma/enums";
 import type { inferSSRProps } from "@calcom/types/inferSSRProps";
 import { Meta, WizardForm } from "@calcom/ui";
 
@@ -18,7 +13,7 @@ import { AdminUserContainer as AdminUser } from "@components/setup/AdminUser";
 import ChooseLicense from "@components/setup/ChooseLicense";
 import EnterpriseLicense from "@components/setup/EnterpriseLicense";
 
-import { ssrInit } from "@server/lib/ssr";
+import { getSetupPageData } from "@server/lib/setupGetData";
 
 function useSetStep() {
   const router = useRouter();
@@ -152,50 +147,4 @@ Setup.isThemeSupported = false;
 Setup.PageWrapper = PageWrapper;
 export default Setup;
 
-export const getServerSideProps = async (context: GetServerSidePropsContext) => {
-  const { req } = context;
-
-  const ssr = await ssrInit(context);
-  const userCount = await prisma.user.count();
-
-  const session = await getServerSession({ req });
-
-  if (session?.user.role && session?.user.role !== UserPermissionRole.ADMIN) {
-    return {
-      redirect: {
-        destination: `/404`,
-        permanent: false,
-      },
-    };
-  }
-
-  const deploymentKey = await prisma.deployment.findUnique({
-    where: { id: 1 },
-    select: { licenseKey: true },
-  });
-
-  // Check existant CALCOM_LICENSE_KEY env var and acccount for it
-  if (!!process.env.CALCOM_LICENSE_KEY && !deploymentKey?.licenseKey) {
-    await prisma.deployment.upsert({
-      where: { id: 1 },
-      update: {
-        licenseKey: process.env.CALCOM_LICENSE_KEY,
-        agreedLicenseAt: new Date(),
-      },
-      create: {
-        licenseKey: process.env.CALCOM_LICENSE_KEY,
-        agreedLicenseAt: new Date(),
-      },
-    });
-  }
-
-  const isFreeLicense = (await getDeploymentKey(prisma)) === "";
-
-  return {
-    props: {
-      trpcState: ssr.dehydrate(),
-      isFreeLicense,
-      userCount,
-    },
-  };
-};
+export const getServerSideProps = getSetupPageData;
